+++ module/module.go
@@ -19,7 +19,12 @@ func init() {
 	Mapper["cli"] = &object.Module{Name: "cli", Functions: CliFunctions}
 	Mapper["term"] = &object.Module{Name: "term", Functions: TermFunctions}
 	Mapper["uuid"] = &object.Module{Name: "uuid", Functions: UuidFunctions}
-	Mapper["string"] = &object.Module{Name: "string", Functions: StringFunctions}
+	Mapper["string"] = &object.Module{
+		Name:        "string",
+		Functions:   StringFunctions,
+		FuncTypes:   StringFuncTypes,
+		FuncReturns: StringFuncReturns,
+	}
 	Mapper["styled"] = &object.Module{Name: "styled", Functions: StyledFunctions}
 	Mapper["crypto"] = &object.Module{Name: "crypto", Functions: CryptoFunctions}
 	Mapper["regex"] = &object.Module{Name: "regex", Functions: RegexFunctions}