+++ token/token.go
@@ -61,7 +61,8 @@ const (
 	RANGE     = ".."
 	ELLIPSIS  = "..."
 	AT        = "@"
-	ARROW     = "=>"
+	ARROW      = "=>"
+	DOUBLECOLON = "::" // builtin function prefix: ::println
 
 	// Keywords
 	FUNCTION = "FUNCTION"
@@ -124,9 +125,9 @@ const (
 	LOG_CAP      = "LOG_CAP"
 
 	// Type system tokens
-	AS   = "AS" // type casting: x as int
-	IS   = "IS" // type checking: x is int
-	PIPE = "|"  // union types: int | string
+	AS   = "AS"   // type casting: x as int
+	IS   = "IS"   // type checking: x is int
+	PIPE = "|"    // union types: int | string
 )
 
 var keywords = map[string]TokenType{